Более-менее отрефакторен код на модули, планирование маршрута заработало, пока что без БД
